+++ CHANGELOG.md
@@ -13,6 +13,23 @@ If your change does not need a CHANGELOG entry, add the "skip changelog" label t
 
 ## Unreleased
 
+- Sign Lambda layer by AWS Signer
+  ([#1275](https://github.com/aws-observability/aws-otel-java-instrumentation/pull/1275))
+- Bump Netty version to 4.1.130 Final
+  ([#1271](https://github.com/aws-observability/aws-otel-java-instrumentation/pull/1271))
+
+
+### Enhancements
+
+- Add Application Signals Dimensions to EMF exporter
+  ([#1264](https://github.com/aws-observability/aws-otel-java-instrumentation/pull/1264))
+- Configure EMF and CompactLog Exporters for Lambda Environment
+  ([#1222](https://github.com/aws-observability/aws-otel-java-instrumentation/pull/1222))
+- feat: [Java] EMF Exporter Implementation
+  ([#1209](https://github.com/aws-observability/aws-otel-java-instrumentation/pull/1209))
+
+## v2.20.0 - 2025-10-29
+
 ### Enhancements
 
 - Support X-Ray Trace Id extraction from Lambda Context object, and respect user-configured OTEL_PROPAGATORS in AWS Lamdba instrumentation